fix: Explicit np.float64 casting in numpy array math and add exception tracking suggestion
- Explicitly cast price history to `np.float64` in `calculate_daily_returns` inside `backend/utils/portfolio_analyzer_additions.py` to fix implicit integer array typecasting errors that crash on math operations.
- Appended a new strategic suggestion (016) regarding centralized error handling to `future/suggestions.MD`.

+++ backend/utils/portfolio_analyzer_additions.py
@@ -10,8 +10,8 @@ class PortfolioAnalyzerAdditions:
     @staticmethod
     def calculate_daily_returns(price_history: List[float], method: str = 'simple') -> np.ndarray:
         if not price_history or len(price_history) < 2:
-            return np.array([])
-        prices = np.array(price_history)
+            return np.array([], dtype=np.float64)
+        prices = np.array(price_history, dtype=np.float64)
         if method == 'log':
             return np.diff(np.log(prices))
         return np.diff(prices) / prices[:-1]
